Removed RecObject.
I want to keep the library minimal. Also, this is not a good way to access
records, since it creates confusion about whether the object is a dictionary
or not. This, for examples doesn't look like a dictionary access:

    rec.name

The user can easily write this class on their own if so desired.